move and refractor all packages and classes under seedu/FlowCLI
